% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_LavaanPlot-package.R
\docType{package}
\name{saveLavaanPlot-package}
\alias{saveLavaanPlot-package}
\title{saveLavaanPlot: Make and save plots with \code{\link{lavaanPlot::lavaanPlot}}}
\description{
Saves a Structural Equation Model (SEM) plot generated by \code{\link{lavaanPlot::lavaanPlot}}. The function handles intermediate steps, including exporting the plot to SVG and converting to PNG.
}
\keyword{internal}
